src/app/services/TodoData/todo-data.service.ts
import { Injectable } from '@angular/core';
import { Todo } from 'src/app/classes/Todo/todo';

@Injectable({
  providedIn: 'root'
})
export class TodoDataService {

  //Automic increment of ids
  lastId: number = 0;

  //Our todos
  todos: Todo[] = [];

  constructor() { }


  // Simulate POST /todos
  addTodo(todo: Todo): TodoDataService {
    if (!todo.id) {
      todo.id = ++this.lastId;
    }
    this.todos.push(todo);
    return this;
  }

  // Simulate DELETE /todos/:id
  deleteTodoById(id: number): TodoDataService {
    this.todos = this.todos
      .filter(todo => todo.id !== id);
    return this;
  }

  // Simulate PUT /todos/:id
  updateTodoById(id: number, values: Object = {}): Todo {
    let todo = this.getTodoById(id);
    if (!todo) {
      return null;
    }
    Object.assign(todo, values);
    return todo;
  }

  // Simulate GET /todos
  getAllTodos(): Todo[] {
    return this.todos;
  }

  getSumAllTodosNotCompleted(): number {
    return this.todos.filter(todo => todo.complete == (false)).length;
  }

  // Simulate GET /todos/:id
  getTodoById(id: number): Todo {
    return this.todos
      .find(todo => todo.id === id);
  }

  // Toggle todo complete
  toggleTodoComplete(todo: Todo){
    let updatedTodo = this.updateTodoById(todo.id, {
      complete: !todo.complete
    });
    return updatedTodo;
  }

}
